requirements.txt
altgraph==0.17.3
Babel==2.12.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
cx-Freeze==6.15.5
cx-Logging==3.1.0
idna==3.4
lief==0.13.2
obs-websocket-py==1.0
pefile==2023.2.7
Pillow==10.0.0
py2exe==0.13.0.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
pywin32-ctypes==0.2.2
reportlab==3.6.8
requests==2.31.0
tkcalendar==1.6.1
urllib3==2.0.3
websocket-client==1.6.1
